anpr/requirements.txt
deep_sort_realtime==1.3.1
numpy==1.23.2
paddleocr==2.6
paddlepaddle==2.3.2
Pillow==9.2.0
yolov7==0.0.1
fastapi[all]
